requirements.txt
django
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
django-environ
django-cors-headers
